refactor(): move sticky nav js in separate file
